Rename client module to follow convention of other client modules
