+++ src/lib/core/hooks/workspace.ts
@@ -1,7 +1,7 @@
 import { useNonNullableContext } from '@veupathdb/wdk-client/lib/Hooks/NonNullableContext';
 import SubsettingClient from '../api/SubsettingClient';
 import DataClient from '../api/DataClient';
-import { AnalysisClient } from '../api/analysis-api';
+import { AnalysisClient } from '../api/AnalysisClient';
 import {
   MakeVariableLink,
   WorkspaceContext,
